requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-auth-httplib2==0.1.1
Pillow==10.2.0
